Refactor: Did some general housekeeping. Added some asserts to classes. Changed some int types to my own typdef'd uint16 or similar. 1 or 2 other sligth modifications.

+++ EngineTester/TestMain.cpp
@@ -1,7 +1,7 @@
 #include "gtest/gtest.h"
 #include "AllEngineTests.h"
 
-int main(int argc, char* argv[])
+uint16 main(int argc, char* argv[])
 {
 	::testing::InitGoogleTest(&argc, argv);
 	RUN_ALL_TESTS();
